backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.2.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
binaryornot==0.4.4
bleach==6.1.0
build==1.0.3
CacheControl==0.13.1
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==2.5.0
crashtest==0.4.1
cryptography==42.0.2
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.1
django-browser-reload==1.12.1
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-jazzmin==2.6.0
django-livereload-server==0.5.1
django-nextjs==2.4.0
django-redis==5.4.0
django-tailwind==3.8.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
frozenlist==1.4.1
idna==3.6
installer==0.7.0
jaraco.classes==3.3.1
Jinja2==3.1.3
keyring==24.3.0
kombu==5.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.2.0
msgpack==1.0.7
multidict==6.0.5
oauthlib==3.2.2
packaging==23.2
pexpect==4.9.0
pillow==10.2.0
pipenv==2023.12.1
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.3
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.0
setuptools==69.1.0
shellingham==1.5.4
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sqlparse==0.4.4
text-unidecode==1.3
tomlkit==0.12.3
tornado==6.4
trove-classifiers==2024.1.31
types-python-dateutil==2.8.19.20240106
tzdata==2024.1
urllib3==2.2.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
xattr==0.10.1
yarl==1.9.4
